Early gradient synchronization with arbitrary language pairs
It is no longer necessary to train all language pairs before synch.

Forward hook to keep track of parameters that have been trained
---------------------------------------------------------------

Using the value of param.grad is not convenient: zero_grad sets it to
zero, which is both difficult to test for and also a valid value for a
gradient.

The hook sets the flag only for parameters that were active during the
call to forward, but not parameters of submodules. It does the right
thing for modules that delegate to submodules conditionally,
e.g. PluggableEmbeddings. It will not do the right thing for modules
that apply parameters conditionally if those parameters are registered
directly to the module itself.

Move scheduling of tasks into Scheduler
---------------------------------------

Corpus ids (language pairs) are sampled from the ids valid on this
device, according to the corpus weights.

Simple curriculum learning
--------------------------

The "from_step" corpus parameter can be used for a simple curriculum:
the corpus id will have zero weight until the specified starting step is
reached.

Option --scheduling_strategy
----------------------------

There are currently two implemented strategies:
- 'sampling' samples tasks according to weight.
- 'roundrobin' goes through tasks in order. Does not support weighting.

Other
-----

- Omit grad if all nodes sent a zero ready bit
- Bugfix: incorrect use of all_reduce return value
- Squashed and ran black to reduce merge conflicts and ease rebase