Plot norms of params and gradients in tensorboard
- Plot L2 norms of params and gradients
- Log losses per task
- Squashed and ran black to reduce merge conflicts and ease rebase

Note: WIP to allow all gpu ranks to log.

Unfortunately: In general, [tensorboard] API expects that
data points are logged with a monotonically increasing step value.
Duplicate points for a single step or points logged out of order by step
are not guaranteed to display as desired in TensorBoard.

This severely limits the ability to log data from different cards
without using a gather op.