Adapters
Inject small bottleneck layers with residual connection into an already
trained network, to adapt it for a new task.

Multiple adapters can be added to the same encoder/decoder. Each adapter
can be individually activated.

Currently all adapters belonging to the same encoder/decoder are
considered to belong the same parameter sharing group as the
encoder/decoder. This is probably not performant enough, if the
encoder/decoder is fully shared.

A mixin is used to create adaptable versions of the Transformer encoder
and decoder. It is not currently possible to adapt other architectures
than the Transformer.

Ids of adapters to activate are included in batch metadata. This allows
activating the correct adapters.

WIP: create_bilingual_model does not yet create adapters. Because of
this, translating is not currently possible.