Only step optimizers of trained modules
Prevent a suboptimizer from being stepped if none of its parameters has a gradient.
If the module is only used in tasks that were not trained since the last
gradient synch, then the optimizer for that module does not need to be
stepped. Stepping updates optimizer parameters, such as gradient moving averages
and per-optimizer warmup.

This means that modules have different effective number of training
steps, depending on the task mix schedule.