Bugfix: Component type without groups
If a component type (e.g. the encoder) only has components trained on a
single gpu, then no communication groups are needed for that component.
This triggered a bug in TaskQueueManager.get_distributed_groups,
where the dict components_to_groups would be missing a key for that
component type.

The unit tests caught another bug in get_distributed_groups, when called
on the global task queue manager or on CPU.

closes #23