Initial implementation of DeviceContext
The global structure is called WorldContext, and the local version with
more fields is called DeviceContext.

Some of the responsibilities of this class were previously in
TaskQueueManager. To reduce the need for refactoring, TQM now delegates
these to WorldContext or DeviceContext as needed.

There is one ugly part, in WorldContext.from_opt, where the number of
slurm nodes is retrieved from ENV. It is not trivial to fix this, as we
1) need the value here
2) maybe don't want to retreive it unless the context is MULTI_GPU
3) determine MULTI_GPU here