n_nodes set in conf, node_id set by srun script
To avoid hard-coding the dependence on slurm, the number of nodes and
the current node are no longer retrieved from slurm-specific environment
variables. Now you must pass them on the command line. The total number
of nodes can be set in the yaml conf, but the main process on each node
must be run with a different value for node_id.