Fix to desync problem
The p.has_grad flag was used in two places:

1) first in only_ready_reduce_and_rescale_grads to determine which
devices have trained a parameter in this minibatch (this usage appears
correct)
2) and later in MultipleOptimizer.step to avoid stepping if
none of the parameters in a module were trained.

The latter didn't work, because the flag had different values on
different devices. This caused devices that didn't train the parameter
in this minibatch to receive the correct gradient from the other
devices, but then refrain from stepping and applying it.

In the current solution, all devices set p.has_grad if the gradient sync
succeeds.