Allow splitting large language pairs in config-config
The goal of splitting LPs is to increase the weight of the LP by running
multiple copies of it in parallel. This avoids problematic task
distributions when a small number of language pairs (fewer than the
number of devices) have much more training data than other LPs.

Because assignment of split LPs on the same GPU would be very
homogenous, a penalty for putting multiple copies of a split LP on a
single device is introduced.