Log the end of training into the structured log file
This allows HPO to detect that training has ended.

The format looks like this:
{"type": "end", "step": 20, "start_time": 1698044256.6410203, "end_time": 1698044361.484478, "duration_s": 104.84345769882202}

Note that currently only a successful end of training is detected: if
the training is killed, e.g. because of a slurm timeout, nothing is
written into the structured log file. If we want to handle these as well
(and we do need to, unless we manage to set the number of training steps
perfectly), we need a signal handler to listen for the signals sent by
slurm. Normally this is SIGTERM sent when the time limit is exceeded,
but it is possible to use the --signal parameter to send other signals
in advance of the time limit. Note that we already have a signal handler
in mammoth.distributed.communication.ErrorHandler

Closes #34