Training with determistic task sampling now works on CPU
- In order to simplify, CPU and single-GPU now also use multiprocessing,
  placing the dataloader in a separate process.

Several places have been refactored to use the new distributed
components:
- Init broadcast in train_single.py
- Gradient communication in trainer.py
  (still uses only_ready_reduce_and_rescale_grads though)
- Sub-optimizer construction in utils/optimizers.py

Two places have been identified as potential future candidates:
- model_builder
- module splitter for checkpointing

The task distribution is now logged in the TQM of the main rank.
It no longer has to be done in the dataloader, as each TQM has a global
view.