Remove special mention of --ff_mult
This flag can be passed though to x-transformers using the
x_transformer_opts dict, there is no need for special handling.
The flag is much less visible now, and a user noticing that
--transformer_ff has been removed will need to dig deep to find it
and learn the changed parameterization.

Not sure where to document this. Maybe we need a new page for migration
tips.